requirements.txt
mysql-connector-python>=8.0.0
pandas>=1.3.0
matplotlib>=3.4.0
python-dotenv>=0.19.0
openai>=1.0.0
flask>=2.0.0
flask-cors>=3.0.0
